+++ changelog.md
@@ -4,6 +4,16 @@ This file contains all the notable changes done to the Ballerina File package th
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- [Change the path separator variables' type from private to public](https://github.com/ballerina-platform/ballerina-standard-library/issues/2831)
+
+## [1.1.0] - 2021-12-13
+
+### Added
+- [Introduce a distinct File service type](https://github.com/ballerina-platform/ballerina-standard-library/issues/2398)
+
 ## [1.1.0-alpha8] - 2021-04-09
 
 ### Changed